refactor: move out lwc test indexer into separate file
