+++ packages/salesforcedx-vscode-lwc/src/testSupport/index.ts
@@ -18,9 +18,5 @@ export function activateLwcTestSupport(context: ExtensionContext) {
   registerLwcTestExplorerTreeView(context);
   startWatchingEditorFocusChange(context);
   taskService.registerTaskService(context);
-  // It's actually a synchronous function to start file watcher.
-  // Finding test files will only happen when going into test explorer
-  // Parsing test files will happen when expanding on the test group nodes,
-  // or open a test file, or on watched files change
-  lwcTestIndexer.configureAndIndex().catch(error => console.error(error));
+  lwcTestIndexer.register(context);
 }
